changekl

+++ toggle.py
@@ -8,7 +8,8 @@ toggle_variable = 0
 GPIO.setmode(GPIO.BOARD)  # Use Broadcom pin-numbering scheme
 GPIO.setup(TOUCH_SENSOR_PIN, GPIO.IN, pull_up_down=GPIO.PUD_UP)
 
-
+bash_process = None
+python_process = None
 
 try:
     while True:
@@ -24,12 +25,22 @@ try:
             else: 
                 print("TERMINATING...")
                 time.sleep(0.5)
-                bash_process.terminate()
-                python_process.terminate()
+                if bash_process:
+                    bash_process.terminate()
+                    bash_process.wait()
+                if python_process:
+                    python_process.terminate()
+                    python_process.wait()
                 toggle_variable = 0
 
             time.sleep(0.5)  
 except KeyboardInterrupt:
     print("Exiting...")
 finally:
+    if bash_process:
+        bash_process.terminate()
+        bash_process.wait()
+    if python_process:
+        python_process.terminate()
+        python_process.wait()
     GPIO.cleanup()
